+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## Next
+>Fix: BezierCircleHeader progress bar is occasionally not hidden  
+>Update: TaurusHeader add backgroundColor parameter   
+
 ## V 2.1.0
 >Fix: After using firstRefreshWidget, pull down refresh does not take effect [issues#250](https://github.com/xuelongqy/flutter_easyrefresh/issues/250),[issues#256](https://github.com/xuelongqy/flutter_easyrefresh/issues/256)